Learning Procedural Mesh Creation and Creating Basic Triangles
Heres the strat:
Because meshes are always created in local space and there is no way to move the object that has a mash without moving the mesh as well, I will implement a non-moving gameobject that gets created when we start our slash. THEN, the gameobject will probably start moving in a direction (The direction the player was looking when they started the swing?) However, the mesh will still be created at our paintbrush, while the object is moving forward, therefore it will give the illusion that the slash is moving forward and we are swinging it. The next step is to figure out the local space/world space transforming I'm going to have to do to get the mesh to be parented to a stable object, but be created at the tip of my paint brush.  Then inputs and actual functionality and stuff.